requirements.txt
crewai
load_dotenv
langchain_google_genai
tools
